move test sources into Tests/ dir
