+++ Diz.Test/Tests/CompressionTest.cs
