+++ Diz.Test/Tests/CpuTests.cs
