+++ Diz.Test/Tests/LogCreator.cs
