+++ Diz.Test/Tests/MiscTests.cs
