+++ Diz.Test/Tests/SerializerDictionaryTest.cs
