+++ Diz.Test/Tests/SuperFamiCheckTests.cs
