+++ Diz.Test/Tests/TraceLogPerformanceTests.cs
