+++ Diz.Test/Tests/TraceLogTests.cs
