+++ Diz.Test/Utils/LogWriterHelper.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using Diz.Core.export;
 using Diz.Core.model;
